Move copyValueListToArrayWriter method to ValueList
