+++ velox/functions/lib/aggregates/ValueList.h
@@ -19,8 +19,11 @@
 #include "velox/common/memory/HashStringAllocator.h"
 #include "velox/exec/Aggregate.h"
 #include "velox/expression/ComplexViewTypes.h"
+#include "velox/expression/ComplexWriterTypes.h"
 #include "velox/vector/ComplexVector.h"
 #include "velox/vector/DecodedVector.h"
+
+using namespace facebook::velox::exec;
 
 namespace facebook::velox::aggregate {
 
@@ -131,5 +134,24 @@ class ValueListReader {
   uint64_t nulls_;
   vector_size_t pos_{0};
 };
+
+// Write ValueList accumulators to Array-typed intermediate or final result
+// vectors.
+// TODO: This API only works if it is the only logic writing to `writer`.
+template <typename T>
+void copyValueListToArrayWriter(ArrayWriter<T>& writer, ValueList& elements) {
+  writer.resetLength();
+  auto size = elements.size();
+  if (size == 0) {
+    return;
+  }
+  writer.reserve(size);
+
+  ValueListReader reader(elements);
+  for (vector_size_t i = 0; i < size; ++i) {
+    reader.next(*writer.elementsVector(), writer.valuesOffset() + i);
+  }
+  writer.resize(size);
+}
 
 } // namespace facebook::velox::aggregate
